Começo de Swagger e sua configuração

+++ src/main/java/br/com/fiap/api_rest/controller/ClienteController.java
@@ -5,6 +5,12 @@ import br.com.fiap.api_rest.dto.ClienteResponse;
 import br.com.fiap.api_rest.model.Cliente;
 import br.com.fiap.api_rest.repository.ClienteRepository;
 import br.com.fiap.api_rest.service.ClienteService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -22,7 +28,8 @@ import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("/clientes")
+@RequestMapping(value = "/clientes", produces = {"application/json"})
+@Tag(name = "api-clientes")
 public class ClienteController {
     @Autowired
     ClienteRepository clienteRepository;
@@ -32,6 +39,14 @@ public class ClienteController {
     //CRUD: CREATE, READ, UPDATE, DELETE - CRUD
     // POST, GET, PUT, DELETE - HTTP CORRESPONDENTES
 
+    @Operation(summary = "Cria um novo cliente")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Cliente cadastrado com sucesso",
+                content = {@Content(mediaType = "application/json",
+                    schema = @Schema(implementation = Cliente.class))}),
+            @ApiResponse(responseCode = "400", description = "Atributos informados são inválidos",
+                        content = @Content(schema = @Schema()))
+    })
     @PostMapping
     public ResponseEntity<Cliente> createCliente(@Valid @RequestBody ClienteRequest cliente){
         Cliente clienteSalvo = clienteRepository.save(clienteService.requestToCliente(cliente));
@@ -39,12 +54,21 @@ public class ClienteController {
     }
 
 
+    @Operation(summary = "Retorna uma lista de clientes")
     @GetMapping
     public ResponseEntity<Page<ClienteResponse>> readClientes(@RequestParam(defaultValue = "0") Integer page) {
         Pageable pageable = PageRequest.of(page, 2, Sort.by("categoria").ascending().and(Sort.by("nome").ascending()));
         return new ResponseEntity<>(clienteService.findAll(pageable), HttpStatus.OK);
     }
 
+    @Operation(summary = "Retorna um cliente por ID")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Cliente encontrado com sucesso",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ClienteResponse.class))}),
+            @ApiResponse(responseCode = "404", description = "Nenhum cliente encontrado",
+                    content = @Content(schema = @Schema()))
+    })
     @GetMapping("/{id}")
     public ResponseEntity<ClienteResponse> readClientes(@PathVariable Long id){
         Optional<Cliente> cliente = clienteRepository.findById(id);
@@ -54,6 +78,14 @@ public class ClienteController {
         return new ResponseEntity<>(clienteService.clienteToResponse(cliente.get(), false), HttpStatus.OK);
     }
 
+    @Operation(summary = "Atualiza um cliente existente")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Cliente encontrado e atualizado com sucesso",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Cliente.class))}),
+            @ApiResponse(responseCode = "400", description = "Nenhum cliente encontrado para atualizar",
+                    content = @Content(schema = @Schema()))
+    })
     @PutMapping("/{id}")
     public ResponseEntity<Cliente> updateCliente(@PathVariable Long id, @RequestBody Cliente cliente){
         Optional<Cliente> clienteExistente = clienteRepository.findById(id);
@@ -65,8 +97,19 @@ public class ClienteController {
         return new ResponseEntity<>(clienteAtualizado, HttpStatus.CREATED);
     }
 
+    @Operation(summary = "Exclui um cliente por ID")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "400", description = "Nenhum cliente encontrado para excluir",
+                    content = @Content(schema = @Schema())),
+            @ApiResponse(responseCode = "204", description = "Cliente excluido com sucesso",
+                    content = @Content(schema = @Schema()))
+    })
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCliente(@PathVariable Long id){
+        Optional<Cliente> clienteExistente = clienteRepository.findById(id);
+        if (clienteExistente.isEmpty()){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         clienteRepository.deleteById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
